requirements.txt
watchdog~=0.10.3
PyYAML~=5.3.1
marshmallow~=3.7.1
python-dotenv~=0.14.0
rootpath~=0.1.1